requirements.txt
Django>=5.0,<6.0
asgiref==3.11.1
crispy-bootstrap5==2025.6
django-ckeditor==6.7.3
django-crispy-forms==2.5
django-js-asset==3.1.2
django-tinymce==5.0.0
pypng==0.20220715.0
sqlparse==0.5.5
typing_extensions==4.15.0